+++ tasks/closureOptions.js
@@ -1,0 +1,39 @@
+/**
+ * Data Objects with various options from the
+ * actual python or java Closure Tools
+ *
+ */
+
+
+var options = {
+  compiler: {}
+};
+
+
+options.compiler.jscomp_off = [
+  'accessControls',
+  'ambiguousFunctionDecl',
+  'checkRegExp',
+  'checkTypes',
+  'checkVars',
+  'constantProperty',
+  'deprecated',
+  'duplicateMessage',
+  'es5Strict',
+  'externsValidation',
+  'fileoverviewTags',
+  'globalThis',
+  'internetExplorerChecks',
+  'invalidCasts',
+  'missingProperties',
+  'nonStandardJsDocs',
+  'strictModuleDepCheck',
+  'typeInvalidation',
+  'undefinedNames',
+  'undefinedVars',
+  'unknownDefines',
+  'uselessCode',
+  'visibility'
+];
+
+module.exports = options;